# Conflicts:
#	software/open-qm-base-station/src/test/java/com/ebp/openQuarterMaster/baseStation/testResources/ui/WebDriverWrapper.java
#	software/open-qm-base-station/src/test/resources/application.yaml

+++ software/open-qm-base-station/src/test/java/com/ebp/openQuarterMaster/baseStation/testResources/ui/WebDriverWrapper.java
@@ -1,14 +1,19 @@
 package com.ebp.openQuarterMaster.baseStation.testResources.ui;
 
 import com.ebp.openQuarterMaster.baseStation.testResources.data.TestUserService;
-import com.ebp.openQuarterMaster.baseStation.testResources.lifecycleManagers.TestResourceLifecycleManager;
 import com.ebp.openQuarterMaster.baseStation.testResources.ui.pages.Root;
+import com.ebp.openQuarterMaster.baseStation.utils.AuthMode;
 import com.ebp.openQuarterMaster.lib.core.user.User;
+import io.github.bonigarcia.wdm.WebDriverManager;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.time.StopWatch;
+import org.eclipse.microprofile.config.ConfigProvider;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.firefox.FirefoxOptions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
@@ -19,6 +24,7 @@ import javax.inject.Inject;
 import java.io.Closeable;
 import java.time.Duration;
 import java.util.List;
+import java.util.concurrent.locks.ReentrantLock;
 
 @Slf4j
 //@RequestScoped
@@ -49,6 +55,10 @@ public class WebDriverWrapper implements Closeable {
 	int port;
 	@ConfigProperty(name = "runningInfo.baseUrl")
 	String baseUrl;
+	@ConfigProperty(name = "service.externalAuth.interactionBase")
+	String keycloakInteractionBase;
+	@ConfigProperty(name = "service.authMode")
+	AuthMode authMode;
 	@Inject
 	TestUserService testUserService;
 	
@@ -60,19 +70,25 @@ public class WebDriverWrapper implements Closeable {
 	@PreDestroy
 	public void close() {
 		log.info("Closing out web driver.");
-		getWebDriver().close();
+		getWebDriver().quit();
 		this.driver = null;
 	}
 	
 	public void cleanup() {
 		log.info("Cleaning up browser after test.");
 		
-		driver.manage().deleteAllCookies();
-		driver.get("about:logo");
-		driver.navigate().refresh();
-		
-//		this.close();
-//		this.setup();
+		if (this.quickClean) {
+			if(AuthMode.EXTERNAL.equals(this.authMode)){
+				driver.get(this.keycloakInteractionBase + "/logout");
+				driver.manage().deleteAllCookies();
+			}
+			driver.manage().deleteAllCookies();
+			driver.get("about:logo");
+			driver.navigate().refresh();
+		} else {
+			this.close();
+			this.setup();
+		}
 	}
 	
 	public WebElement findElement(By by) {
